requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
laspy==2.6.1
lazrs==0.7.0
numpy==2.2.6
sqlparse==0.5.3
typing_extensions==4.14.1
